Rename DTO namespace to ValueObjects to match Laravel documentation
https://laravel.com/docs/9.x/eloquent-mutators#value-object-casting
